Move font-awesome submodule to a module directory
By having font-awesome directly in static, the src directory and other
unneeded files would get pulled with it to the public directory. Moving it
to a module directory and symlinking what's needed appears to be the
best solution.

At first I tried to do the same thing using dirlinks, but apparently
hugo doesn't support dirlinks for whatever reason. So symlinks for each
file is the only choice left.