Update config file for screenshot
This commit updates the config file for screenshot. The old config file
uses obsolete features and has worse documentation than the current
default config file. The updated config file is based off of the default
config file of dunst v1.7.3 with some options changed to match the
behavior of the old config file.

This commit also marks on the screenshot page that the config is
updated, so that users will notice that the file has changed.